Made requested changes from PR comments.
