+++ src/app/globals.css
@@ -3,15 +3,15 @@
 @tailwind utilities;
 
 .main {
-  @apply h-screen bg-[url('/background.png')] mt-[-80px] bg-cover bg-no-repeat flex flex-col gap-16 justify-center items-center text-3xl md:text-5xl text-white;
+  @apply min-h-screen bg-[url('/background.png')] -mt-20 bg-cover bg-no-repeat flex flex-col gap-16 justify-center items-center text-3xl md:text-5xl text-white;
 }
 
 .nav {
-  @apply w-full h-[80px] absolute top-0 px-[50px] flex flex-row items-center justify-between z-[1000] text-xl;
+  @apply w-full h-20 absolute top-0 md:px-6 lg:px-12 flex flex-row items-center justify-between z-[1000] text-xl;
 }
 
 p {
-  @apply py-3 text-left leading-8;
+  @apply md:py-1 lg:py-3 text-left leading-8;
 }
 
 .btn-nav {
