+++ src/components/navigation/constants.ts
@@ -1,0 +1,6 @@
+export const navLinks = [
+  { href: "/about", name: "About" },
+  { href: "/shop", name: "Shop" },
+  { href: "/contact", name: "Contact" },
+  { href: "/cart", name: "Cart" },
+];
